move collapse into expand
